feat: reorganize e2e tests for SSE and HTTP streaming servers - rename httpstream-server.e2e.spec.ts to sse-server.e2e.spec.ts for legacy server testing - create new httpstream-server.e2e.spec.ts for modern HTTP streaming server testing - add comprehensive session management tests for modern server - update documentation with test information and environment variables - separate test ports to avoid conflicts (3001 for modern, 3002 for legacy)
