+++ src/tests/e2e/httpstream-server.e2e.spec.ts
@@ -1,9 +1,9 @@
-import request from 'supertest'; // For HTTP requests
+import request from 'supertest';
 import { ChildProcess, spawn } from 'child_process';
 import path from 'path';
-import { MEMORY_BANK_MCP_TOOLS } from '../../mcp/tools'; // Adjust path as needed
-import { Component, ComponentStatus } from '../../types'; // Adjust path
-import { setupTestDB, cleanupTestDB } from '../utils/test-db-setup'; // Removed dbPath import
+import { MEMORY_BANK_MCP_TOOLS } from '../../mcp/tools';
+import { Component, ComponentStatus } from '../../types';
+import { setupTestDB, cleanupTestDB } from '../utils/test-db-setup';
 
 // Increase Jest timeout
 jest.setTimeout(90000);
@@ -12,216 +12,23 @@ const STREAM_PORT = process.env.HTTP_STREAM_PORT || 3001;
 const STREAM_HOST = process.env.HOST || 'localhost';
 const BASE_URL = `http://${STREAM_HOST}:${STREAM_PORT}`;
 
-// Interface for the resolved value of collectStreamEvents
-interface CollectedSseEvents {
-  events: any[];
-  progressEventsCount: number;
-  finalResponseEvent: any | null;
-  errorEvent: any | null;
-  connectionError?: Error;
-}
-
-// Copied from httpstream-streaming-implementation.md guide
-const collectStreamEvents = async (sseResponseEmitter: any): Promise<CollectedSseEvents> => {
-  const events: any[] = [];
-  let progressEventsCount = 0;
-  let finalResponseEvent: any | null = null;
-  let errorEvent: any | null = null;
-
-  console.log('[collectStreamEvents] Starting collection of SSE events...');
-
-  return new Promise<CollectedSseEvents>((resolve, reject) => {
-    let promiseSettled = false;
-
-    const settlePromise = (
-      resolver: (value: CollectedSseEvents | PromiseLike<CollectedSseEvents>) => void,
-      value: CollectedSseEvents,
-      isRejectOperation = false,
-    ) => {
-      if (!promiseSettled) {
-        promiseSettled = true;
-        console.log(
-          `[collectStreamEvents] Promise being ${isRejectOperation ? 'rejected' : 'resolved'}.`,
-        );
-        if (sseResponseEmitter.destroy) {
-          console.log('[collectStreamEvents] Destroying SSE response emitter.');
-          sseResponseEmitter.destroy();
-        } else if (sseResponseEmitter.removeAllListeners) {
-          console.log('[collectStreamEvents] Removing all listeners from SSE response emitter.');
-          sseResponseEmitter.removeAllListeners();
-        }
-        resolver(value);
-      }
-    };
-
-    let buffer = '';
-    const processBuffer = () => {
-      if (promiseSettled) {
-        return;
-      }
-
-      let eolIndex;
-      while ((eolIndex = buffer.indexOf('\n\n')) >= 0) {
-        if (promiseSettled) {
-          break;
-        }
-        const message = buffer.substring(0, eolIndex);
-        buffer = buffer.substring(eolIndex + 2);
-        if (message.trim() === '') {
-          continue;
-        }
-
-        const lines = message.split('\n');
-        let eventType = 'message';
-        let eventDataString = '';
-        let eventId = null;
-
-        lines.forEach((line) => {
-          if (line.startsWith('event:')) {
-            eventType = line.substring('event:'.length).trim();
-          } else if (line.startsWith('data:')) {
-            eventDataString = line.substring('data:'.length).trim();
-          } else if (line.startsWith('id:')) {
-            eventId = line.substring('id:'.length).trim();
-          }
-        });
-
-        try {
-          let parsedData: any = {}; // Default to empty object, typed as any
-          if (eventDataString) {
-            // Only parse if eventDataString is not empty
-            parsedData = JSON.parse(eventDataString);
-          }
-          const eventPayload = { type: eventType, id: eventId, data: parsedData };
-          events.push(eventPayload);
-          console.log(`[collectStreamEvents] Received event:`, eventPayload);
-
-          if (eventType === 'mcpNotification' && parsedData.method === 'tools/progress') {
-            progressEventsCount++;
-            console.log(`[collectStreamEvents] Progress event count: ${progressEventsCount}`);
-          }
-          if (eventType === 'mcpResponse') {
-            finalResponseEvent = eventPayload;
-            console.log(
-              '[collectStreamEvents] Final mcpResponse event received. Resolving promise.',
-            );
-            settlePromise(resolve, { events, progressEventsCount, finalResponseEvent, errorEvent });
-            return;
-          }
-          if (eventType === 'error') {
-            errorEvent = eventPayload;
-            console.error(
-              '[collectStreamEvents] SSE standard error event received. Rejecting promise:',
-              eventPayload,
-            );
-            const syntheticError = new Error(parsedData.message || 'SSE error event');
-            settlePromise(
-              reject,
-              {
-                events,
-                progressEventsCount,
-                finalResponseEvent,
-                errorEvent,
-                connectionError: syntheticError,
-              },
-              true,
-            );
-            return;
-          }
-        } catch (e) {
-          console.error(
-            '[collectStreamEvents] Failed to parse SSE event data:',
-            eventDataString,
-            e,
-          );
-          // Potentially reject here if parsing error is critical, for now, it logs and continues.
-          // If a parsing error should stop everything:
-          // errorEvent = { type: 'parsing_error', data: { message: (e as Error).message, rawData: eventDataString }};
-          // settlePromise(reject, { events, progressEventsCount, finalResponseEvent, errorEvent, connectionError: e as Error }, true);
-          // return;
-        }
-      }
-    };
-
-    if (sseResponseEmitter.on) {
-      sseResponseEmitter.on('data', (chunk: Buffer | string) => {
-        if (promiseSettled) {
-          return;
-        }
-        buffer += chunk.toString();
-        console.log(`[collectStreamEvents] Data chunk received (${chunk.length} bytes)`);
-        processBuffer();
-      });
-      sseResponseEmitter.on('end', () => {
-        if (promiseSettled) {
-          return;
-        }
-        console.log('[collectStreamEvents] SSE stream "end" event received.');
-        if (buffer.length > 0) {
-          processBuffer(); // Process any remaining buffer
-        }
-        if (promiseSettled) {
-          return;
-        } // Check if processBuffer settled it
-        // If stream ends and we haven't resolved (e.g. no mcpResponse), resolve with current state.
-        // This maintains original behavior for streams that might end without explicit mcpResponse.
-        console.log(
-          '[collectStreamEvents] Resolving promise due to stream "end" event (if not already settled).',
-        );
-        settlePromise(resolve, { events, progressEventsCount, finalResponseEvent, errorEvent });
-      });
-      sseResponseEmitter.on('error', (err: Error) => {
-        // Connection/transport error
-        if (promiseSettled) {
-          return;
-        }
-        console.error('[collectStreamEvents] SSE stream connection error. Rejecting promise:', err);
-        if (!errorEvent) {
-          errorEvent = { type: 'connection_error', data: { message: err.message } };
-        }
-        settlePromise(
-          reject,
-          { events, progressEventsCount, finalResponseEvent, errorEvent, connectionError: err },
-          true,
-        );
-      });
-    } else {
-      console.error(
-        '[collectStreamEvents] Provided sseResponseEmitter does not have .on method for events',
-      );
-      const err = new Error('Provided sseResponseEmitter does not have .on method for events');
-      settlePromise(
-        reject,
-        {
-          events: [],
-          progressEventsCount: 0,
-          finalResponseEvent: null,
-          errorEvent: { type: 'setup_error', data: { message: err.message } },
-          connectionError: err,
-        },
-        true,
-      );
-    }
-  });
-};
-
-describe('MCP HTTP Stream Server E2E Tests', () => {
+describe('MCP HTTP Streaming Server E2E Tests (Modern)', () => {
   let serverProcess: ChildProcess;
-  let dbPathForTest: string; // Will be set by setupTestDB return value
-  let clientProjectRootForTest: string; // <<<< Dependant on dbPathForTest
+  let dbPathForTest: string;
+  let clientProjectRootForTest: string;
   let mcpSessionId: string; // Store the MCP session ID for subsequent requests
-  const testRepository = 'e2e-httpstream-repo'; // Specific repo for these tests
+  const testRepository = 'e2e-httpstream-modern-repo';
   const testBranch = 'main';
   let testComponentId: string | null = null;
   let dependentComponentId: string;
 
   // Share these across describe blocks if tests depend on IDs created in earlier blocks
   let sharedTestComponentId: string | null = null;
-  let sharedDependentComponentId: string | null = null;
-  let sharedTestDecisionId: string | null = null;
-  let sharedTestRuleId: string | null = null;
+  const sharedDependentComponentId: string | null = null;
+  const sharedTestDecisionId: string | null = null;
+  const sharedTestRuleId: string | null = null;
 
-  // Helper function to make MCP requests with proper session management and SSE handling
+  // Helper function to make MCP requests with proper session management
   const makeMcpRequest = async (payload: any, expectStatus = 200) => {
     const headers: any = {
       'Content-Type': 'application/json',
@@ -233,64 +40,23 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
       headers['mcp-session-id'] = mcpSessionId;
     }
 
-    return new Promise<any>((resolve, reject) => {
-      const sseRequest = request(BASE_URL).post('/mcp').set(headers).send(payload);
+    const response = await request(BASE_URL)
+      .post('/mcp')
+      .set(headers)
+      .send(payload)
+      .expect(expectStatus);
 
-      sseRequest
-        .expect(expectStatus)
-        .parse(async (res: any, callback: any) => {
-          try {
-            const { events, finalResponseEvent, errorEvent }: CollectedSseEvents =
-              await collectStreamEvents(res);
-
-            if (errorEvent) {
-              reject(new Error(`SSE error: ${JSON.stringify(errorEvent)}`));
-              return;
-            }
-
-            // Look for the response event with the matching request ID
-            const responseEvent = events.find(
-              (event: any) =>
-                event.type === 'message' &&
-                event.data &&
-                event.data.id === payload.id &&
-                (event.data.result || event.data.error),
-            );
-
-            if (!responseEvent) {
-              reject(new Error(`No response event found for request ID: ${payload.id}`));
-              return;
-            }
-
-            // Return a response object that mimics the old format
-            const response = {
-              body: responseEvent.data,
-              headers: res.headers,
-            };
-
-            callback(null, null);
-            resolve(response);
-          } catch (error) {
-            callback(error, null);
-            reject(error);
-          }
-        })
-        .end((err: any) => {
-          if (err) {
-            reject(err);
-          }
-        });
-    });
+    return response;
   };
 
-  // Helper function to initialize MCP session with SSE handling
+  // Helper function to initialize MCP session
   const initializeMcpSession = async () => {
     const initializePayload = {
       jsonrpc: '2.0',
       id: 'initialize',
       method: 'initialize',
       params: {
-        protocolVersion: '0.1',
+        protocolVersion: '2025-03-26',
         capabilities: {
           roots: {
             listChanged: true,
@@ -304,100 +70,56 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
       },
     };
 
-    return new Promise<void>((resolve, reject) => {
-      const sseRequest = request(BASE_URL)
-        .post('/mcp')
-        .set({
-          'Content-Type': 'application/json',
-          Accept: 'application/json, text/event-stream',
-          Origin: 'http://localhost',
-        })
-        .send(initializePayload);
+    const response = await request(BASE_URL)
+      .post('/mcp')
+      .set({
+        'Content-Type': 'application/json',
+        Accept: 'application/json, text/event-stream',
+        Origin: 'http://localhost',
+      })
+      .send(initializePayload)
+      .expect(200);
 
-      sseRequest
-        .expect(200)
-        .parse(async (res: any, callback: any) => {
-          try {
-            // Extract session ID from headers
-            mcpSessionId = res.headers['mcp-session-id'];
-            console.log(`MCP session ID received: ${mcpSessionId}`);
+    // Extract session ID from headers
+    mcpSessionId = response.headers['mcp-session-id'];
+    console.log(`MCP session ID received: ${mcpSessionId}`);
 
-            const { events, finalResponseEvent, errorEvent }: CollectedSseEvents =
-              await collectStreamEvents(res);
+    // Validate the initialize response
+    expect(response.body.result).toBeDefined();
+    expect(response.body.result.capabilities).toBeDefined();
+    expect(response.body.result.serverInfo.name).toBe('KuzuMem-MCP-HTTPStream');
+    expect(mcpSessionId).toBeDefined();
 
-            if (errorEvent) {
-              reject(new Error(`SSE error: ${JSON.stringify(errorEvent)}`));
-              return;
-            }
-
-            // For initialize, look for the message event with id 'initialize'
-            const initializeEvent = events.find(
-              (event: any) =>
-                event.type === 'message' &&
-                event.data &&
-                event.data.id === 'initialize' &&
-                event.data.result,
-            );
-
-            if (!initializeEvent) {
-              reject(new Error('No initialize response event received'));
-              return;
-            }
-
-            console.log('Initialize response event:', JSON.stringify(initializeEvent, null, 2));
-
-            // Validate the initialize response
-            expect(initializeEvent.data.result).toBeDefined();
-            expect(initializeEvent.data.result.capabilities.tools).toBeDefined();
-            expect(initializeEvent.data.result.serverInfo.name).toBe('KuzuMem-MCP-HTTPStream');
-            expect(mcpSessionId).toBeDefined();
-
-            console.log(`MCP session initialized successfully: ${mcpSessionId}`);
-            callback(null, null);
-            resolve();
-          } catch (error) {
-            callback(error, null);
-            reject(error);
-          }
-        })
-        .end((err: any) => {
-          if (err) {
-            reject(err);
-          }
-        });
-    });
+    console.log(`MCP session initialized successfully: ${mcpSessionId}`);
   };
 
   const startHttpStreamServer = (envVars: Record<string, string> = {}): Promise<void> => {
     return new Promise((resolve, reject) => {
       const serverFilePath = path.resolve(__dirname, '../../mcp-httpstream-server.ts');
-      // Use STREAM_PORT for the PORT env var for this server
       const defaultEnv = {
         PORT: String(STREAM_PORT),
-        DEBUG: '3', // Increase debug level for more verbose logging
+        DEBUG_LEVEL: '3',
+        SESSION_TIMEOUT: '300000', // 5 minutes for tests
         DB_FILENAME: dbPathForTest,
         ...envVars,
       };
 
-      console.log('Starting HTTP Stream server from path:', serverFilePath);
+      console.log('Starting HTTP Streaming server from path:', serverFilePath);
       console.log('With environment variables:', JSON.stringify(defaultEnv, null, 2));
 
       serverProcess = spawn('npx', ['ts-node', '--transpile-only', serverFilePath], {
         env: { ...process.env, ...defaultEnv },
         shell: true,
-        detached: false, // Changed to false for simpler process killing
+        detached: false,
       });
 
       let output = '';
       let resolved = false;
       let startupTimeout: NodeJS.Timeout | null = null;
 
-      // Set a timeout to reject the promise if the server doesn't start within 30 seconds
       startupTimeout = setTimeout(() => {
         if (!resolved) {
-          console.error(
-            `HTTP Stream Server failed to start within timeout. Output so far:\n${output}`,
-          );
+          console.error(`HTTP Streaming Server failed to start within timeout. Output:\n${output}`);
           resolved = true;
           reject(new Error('Server startup timeout'));
         }
@@ -406,14 +128,14 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
       serverProcess.stdout?.on('data', (data) => {
         const sData = data.toString();
         output += sData;
-        console.log(`HTTPSTREAM_SERVER_STDOUT: ${sData}`);
+        console.log(`HTTP_STREAM_SERVER_STDOUT: ${sData}`);
         if (
           !resolved &&
           sData.includes(
-            `MCP HTTP Streaming Server running at http://${STREAM_HOST}:${STREAM_PORT}`,
+            `MCP HTTP Streaming Server (v3.0.0) running at http://${STREAM_HOST}:${STREAM_PORT}`,
           )
         ) {
-          console.log('HTTP Stream Server ready.');
+          console.log('HTTP Streaming Server ready.');
           if (startupTimeout) {
             clearTimeout(startupTimeout);
           }
@@ -421,13 +143,13 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
           resolve();
         }
       });
+
       serverProcess.stderr?.on('data', (data) => {
         const sData = data.toString();
         output += sData;
-        console.error(`HTTPSTREAM_SERVER_STDERR: ${sData}`);
+        console.error(`HTTP_STREAM_SERVER_STDERR: ${sData}`);
         if (!resolved && (sData.includes('EADDRINUSE') || sData.toLowerCase().includes('error'))) {
-          // If an error occurs before resolving, reject to prevent test timeout
-          console.error('HTTP Stream Server failed to start due to an error');
+          console.error('HTTP Streaming Server failed to start due to an error');
           if (startupTimeout) {
             clearTimeout(startupTimeout);
           }
@@ -435,8 +157,9 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
           reject(new Error(`Server failed to start: ${sData.substring(0, 300)}`));
         }
       });
+
       serverProcess.on('error', (err) => {
-        console.error('Failed to start HTTP Stream server process:', err);
+        console.error('Failed to start HTTP Streaming server process:', err);
         if (!resolved) {
           if (startupTimeout) {
             clearTimeout(startupTimeout);
@@ -445,10 +168,11 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
           reject(err);
         }
       });
+
       serverProcess.on('exit', (code, signal) => {
         if (code !== 0 && signal !== 'SIGTERM' && signal !== 'SIGINT') {
           console.warn(
-            `HTTP Stream Server process exited unexpectedly with code ${code}, signal ${signal}. Output:\n${output}`,
+            `HTTP Streaming Server process exited unexpectedly with code ${code}, signal ${signal}. Output:\n${output}`,
           );
           if (!resolved) {
             if (startupTimeout) {
@@ -463,7 +187,7 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
   };
 
   beforeAll(async () => {
-    dbPathForTest = await setupTestDB('httpstream_e2e_test.kuzu');
+    dbPathForTest = await setupTestDB('httpstream_modern_e2e_test.kuzu');
     clientProjectRootForTest = path.dirname(dbPathForTest);
     await startHttpStreamServer();
 
@@ -473,65 +197,54 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
     const repository = testRepository;
     const branch = testBranch;
 
-    console.log(`HTTP Stream E2E: Initializing memory bank for ${repository}...`);
+    console.log(`HTTP Streaming E2E: Initializing memory bank for ${repository}...`);
     const initPayload = {
       jsonrpc: '2.0',
-      id: 'init_e2e_http',
+      id: 'init_e2e_http_modern',
       method: 'tools/call',
       params: {
         name: 'init-memory-bank',
         arguments: { repository, branch, clientProjectRoot: clientProjectRootForTest },
       },
     };
+
     let httpResponse = await makeMcpRequest(initPayload);
-    expect(httpResponse.body.id).toBe('init_e2e_http');
+    expect(httpResponse.body.id).toBe('init_e2e_http_modern');
     expect(httpResponse.body.result?.content).toBeDefined();
     expect(httpResponse.body.result?.content[0]).toBeDefined();
-    
+
     // Parse the JSON result from the MCP SDK format
     const initResult = JSON.parse(httpResponse.body.result.content[0].text);
     expect(initResult.success).toBe(true);
-    console.log('HTTP Stream E2E: Memory bank initialized.');
+    console.log('HTTP Streaming E2E: Memory bank initialized.');
 
-    // --- BEGIN FULL DATA SEEDING (adapted from stdio tests) ---
-    console.log('HTTP Stream E2E: Seeding database with initial data...');
+    // Seed test data
+    console.log('HTTP Streaming E2E: Seeding database with initial data...');
 
     // Seed Components
     const componentsToSeed = [
       {
-        id: 'comp-seed-001',
-        name: 'Seeded Component Alpha',
+        id: 'comp-modern-001',
+        name: 'Modern Component Alpha',
         kind: 'library',
         status: 'active' as ComponentStatus,
       },
       {
-        id: 'comp-seed-002',
-        name: 'Seeded Component Beta',
+        id: 'comp-modern-002',
+        name: 'Modern Component Beta',
         kind: 'service',
         status: 'active' as ComponentStatus,
-        depends_on: ['comp-seed-001'],
+        depends_on: ['comp-modern-001'],
       },
       {
-        id: 'comp-seed-003',
-        name: 'Seeded Component Gamma',
+        id: 'comp-modern-003',
+        name: 'Modern Component Gamma',
         kind: 'API',
         status: 'planned' as ComponentStatus,
       },
-      {
-        id: 'comp-seed-004',
-        name: 'Seeded Component Delta',
-        kind: 'database',
-        status: 'deprecated' as ComponentStatus,
-      },
-      {
-        id: 'comp-seed-005',
-        name: 'Seeded Component Epsilon',
-        kind: 'UI',
-        status: 'active' as ComponentStatus,
-      },
     ];
-    testComponentId = componentsToSeed[0].id; // Save one for later tests if needed
-    dependentComponentId = componentsToSeed[1].id; // Save one for later tests if needed
+    testComponentId = componentsToSeed[0].id;
+    dependentComponentId = componentsToSeed[1].id;
 
     for (const comp of componentsToSeed) {
       const addCompPayload = {
@@ -548,21 +261,15 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
       const compResult = JSON.parse(httpResponse.body.result.content[0].text);
       expect(compResult.success).toBe(true);
     }
-    console.log(`${componentsToSeed.length} components seeded for HTTP E2E.`);
+    console.log(`${componentsToSeed.length} components seeded for HTTP Streaming E2E.`);
 
     // Seed Contexts
     const contextsToSeed = [
       {
-        summary: 'Initial context for seeding via HTTP',
-        agent: 'seed-script-http',
-        decisions: ['DEC-SEED-001'],
-        observations: ['OBS-SEED-001'],
-      },
-      {
-        summary: 'Another seeded context entry via HTTP',
-        agent: 'seed-script-http',
-        decisions: ['DEC-SEED-002'],
-        observations: ['OBS-SEED-002', 'OBS-SEED-003'],
+        summary: 'Initial context for modern HTTP streaming',
+        agent: 'test-agent-modern',
+        decisions: ['DEC-MODERN-001'],
+        observations: ['OBS-MODERN-001'],
       },
     ];
     for (const ctxData of contextsToSeed) {
@@ -585,138 +292,83 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
       const ctxResult = JSON.parse(httpResponse.body.result.content[0].text);
       expect(ctxResult.success).toBe(true);
     }
-    console.log(`${contextsToSeed.length} context entries updated/seeded for HTTP E2E.`);
-
-    // Seed Decisions
-    const decisionsToSeed = [
-      {
-        id: 'dec-seed-001',
-        name: 'Seeded HTTP Decision Alpha',
-        date: '2023-01-01',
-        context: 'Regarding initial HTTP setup',
-      },
-      {
-        id: 'dec-seed-002',
-        name: 'Seeded HTTP Decision Beta',
-        date: '2023-01-15',
-        context: 'HTTP Architectural choice',
-      },
-    ];
-    for (const dec of decisionsToSeed) {
-      const addDecPayload = {
-        jsonrpc: '2.0',
-        id: `add_dec_${dec.id}`,
-        method: 'tools/call',
-        params: {
-          name: 'add-decision',
-          arguments: { repository, branch, ...dec, clientProjectRoot: clientProjectRootForTest },
-        },
-      };
-      httpResponse = await makeMcpRequest(addDecPayload);
-      expect(httpResponse.body.result?.content).toBeDefined();
-      const decResult = JSON.parse(httpResponse.body.result.content[0].text);
-      expect(decResult.success).toBe(true);
-    }
-    console.log(`${decisionsToSeed.length} decisions seeded for HTTP E2E.`);
-
-    // Seed Rules
-    const rulesToSeed = [
-      {
-        id: 'rule-seed-001',
-        name: 'Seeded HTTP Rule Alpha',
-        created: '2023-01-05',
-        content: 'Standard HTTP linting',
-        status: 'active' as const,
-      },
-      {
-        id: 'rule-seed-002',
-        name: 'Seeded HTTP Rule Beta',
-        created: '2023-01-20',
-        content: 'HTTP Security check',
-        status: 'active' as const,
-        triggers: ['commit', 'push'],
-      },
-    ];
-    for (const rule of rulesToSeed) {
-      const addRulePayload = {
-        jsonrpc: '2.0',
-        id: `add_rule_${rule.id}`,
-        method: 'tools/call',
-        params: {
-          name: 'add-rule',
-          arguments: { repository, branch, ...rule, clientProjectRoot: clientProjectRootForTest },
-        },
-      };
-      httpResponse = await makeMcpRequest(addRulePayload);
-      expect(httpResponse.body.result?.content).toBeDefined();
-      const ruleResult = JSON.parse(httpResponse.body.result.content[0].text);
-      expect(ruleResult.success).toBe(true);
-    }
-    console.log(`${rulesToSeed.length} rules seeded for HTTP E2E.`);
-    console.log('HTTP Stream E2E: Database seeding complete.');
-    // --- END DATA SEEDING ---
+    console.log('HTTP Streaming E2E: Database seeding complete.');
   }, 120000);
 
   afterAll(async () => {
+    // Terminate session if it exists
+    if (mcpSessionId) {
+      try {
+        await request(BASE_URL).delete('/mcp').set('mcp-session-id', mcpSessionId).expect(200);
+        console.log('MCP session terminated successfully');
+      } catch (error) {
+        console.warn('Failed to terminate MCP session:', error);
+      }
+    }
+
     if (serverProcess && serverProcess.pid && !serverProcess.killed) {
-      console.log('Stopping HTTP Stream server...');
+      console.log('Stopping HTTP Streaming server...');
       serverProcess.kill();
       await new Promise((resolve) => setTimeout(resolve, 1000));
     }
-    // Pass the specific path to cleanupTestDB
+
     if (dbPathForTest) {
       await cleanupTestDB(dbPathForTest);
     }
   });
 
-  it('T_HTTPSTREAM_001: /initialize should return server capabilities', async () => {
-    // This test verifies that the session was properly initialized in beforeAll
+  it('T_HTTP_MODERN_001: Should initialize and manage MCP session', async () => {
     expect(mcpSessionId).toBeDefined();
-    console.log(`Test T_HTTPSTREAM_001: Using session ${mcpSessionId}`);
-    
-    // Session should already be initialized, so we can verify it's working by making a simple request
+    console.log(`Test T_HTTP_MODERN_001: Using session ${mcpSessionId}`);
+
+    // Verify session is working with a simple request
     const response = await request(BASE_URL)
-      .get('/tools/list')
+      .get('/health')
       .set('Origin', 'http://localhost')
       .expect('Content-Type', /json/)
       .expect(200);
 
-    expect(Array.isArray(response.body.tools)).toBe(true);
-    expect(response.body.tools.length).toBe(MEMORY_BANK_MCP_TOOLS.length);
-    expect(response.body.tools[0]).toHaveProperty('name');
-    expect(response.body.tools[0]).toHaveProperty('inputSchema');
+    expect(response.body.status).toBe('healthy');
+    expect(response.body.transport).toBe('streamable-http');
+    expect(response.body.sessions).toBeGreaterThanOrEqual(1); // At least our session
+    expect(response.body.version).toBe('3.0.0');
   });
 
-  it('T_HTTPSTREAM_002: /tools/list should return list of tools', async () => {
-    const response = await request(BASE_URL)
-      .get('/tools/list')
-      .set('Origin', 'http://localhost')
-      .expect('Content-Type', /json/)
-      .expect(200);
+  it('T_HTTP_MODERN_002: Should list tools via tools/list', async () => {
+    const listToolsPayload = {
+      jsonrpc: '2.0',
+      id: 'list_tools_modern',
+      method: 'tools/list',
+      params: {},
+    };
 
-    expect(Array.isArray(response.body.tools)).toBe(true);
-    expect(response.body.tools.length).toBe(MEMORY_BANK_MCP_TOOLS.length);
-    expect(response.body.tools[0]).toHaveProperty('name');
-    expect(response.body.tools[0]).toHaveProperty('inputSchema');
+    const response = await makeMcpRequest(listToolsPayload);
+    expect(response.body.id).toBe('list_tools_modern');
+    expect(response.body.result).toBeDefined();
+    expect(response.body.result.tools).toBeDefined();
+    expect(Array.isArray(response.body.result.tools)).toBe(true);
+    expect(response.body.result.tools.length).toBe(MEMORY_BANK_MCP_TOOLS.length);
+    expect(response.body.result.tools[0]).toHaveProperty('name');
+    expect(response.body.result.tools[0]).toHaveProperty('inputSchema');
   });
 
-  it('T_HTTPSTREAM_003: /mcp tools/call (JSON) should execute a batch tool', async () => {
-    const repoName = 'e2e-httpstream-repo';
+  it('T_HTTP_MODERN_003: Should execute tools/call with session context', async () => {
+    const repoName = testRepository;
     const metadataContent = {
-      id: 'meta',
+      id: 'meta-modern',
       project: {
         name: repoName,
         created: new Date().toISOString().split('T')[0],
-        description: 'Test E2E HTTP',
+        description: 'Modern HTTP Streaming Test',
       },
       tech_stack: { language: 'TypeScript' },
-      architecture: 'test-driven',
+      architecture: 'microservices',
       memory_spec_version: '3.0.0',
     };
-    
+
     const payload = {
       jsonrpc: '2.0',
-      id: 'update_meta_http',
+      id: 'update_meta_modern',
       method: 'tools/call',
       params: {
         name: 'update-metadata',
@@ -730,441 +382,139 @@ describe('MCP HTTP Stream Server E2E Tests', () => {
     };
 
     const response = await makeMcpRequest(payload);
-    expect(response.body.id).toBe('update_meta_http');
+    expect(response.body.id).toBe('update_meta_modern');
     expect(response.body.result).toBeDefined();
     expect(response.body.result.content).toBeDefined();
     expect(response.body.result.content[0]).toBeDefined();
 
-    // Parse the JSON result from the MCP SDK format
     const toolResult = JSON.parse(response.body.result.content[0].text);
     expect(toolResult.success).toBe(true);
   });
 
-  it.skip('T_HTTPSTREAM_004: /mcp tools/call (SSE) should stream progress for get-component-dependencies', (done) => {
-    // Ensure dependentComponentId and testComponentId are seeded from beforeAll
-    expect(dependentComponentId).toBeDefined();
-    expect(testComponentId).toBeDefined();
+  it('T_HTTP_MODERN_004: Should handle session termination', async () => {
     expect(mcpSessionId).toBeDefined();
 
-    const toolArgs = {
-      repository: testRepository,
-      branch: testBranch,
-      componentId: dependentComponentId,
-      depth: 1, // Keep depth 1 for simpler expected output for now
-      clientProjectRoot: clientProjectRootForTest,
-    };
-    const payload = {
-      jsonrpc: '2.0',
-      id: 'sse_get_deps_http',
-      method: 'tools/call',
-      params: { name: 'get-component-dependencies', arguments: toolArgs },
-    };
+    // Test that we can terminate the session
+    const response = await request(BASE_URL)
+      .delete('/mcp')
+      .set('mcp-session-id', mcpSessionId)
+      .expect(200);
 
-    const sseRequest = request(BASE_URL)
+    expect(response.text).toBe('Session terminated');
+
+    // After termination, requests with the old session ID should fail
+    const testResponse = await request(BASE_URL)
       .post('/mcp')
-      .set('Origin', 'http://localhost')
-      .set('mcp-session-id', mcpSessionId) // Use the session ID
-      .set('Accept', 'text/event-stream')
-      .send(payload);
-
-    // Note: The official MCP SDK may not support SSE streaming for HTTP transport
-    // This test may need to be updated based on the actual SDK capabilities
-    sseRequest
-      .expect(200) // Expect initial 200 OK for SSE stream connection
-      .parse(async (res: any, callback: any) => {
-        try {
-          const {
-            events,
-            progressEventsCount,
-            finalResponseEvent,
-            errorEvent,
-          }: CollectedSseEvents = await collectStreamEvents(res);
-
-          // Basic checks
-          expect(errorEvent).toBeNull(); // Should be no protocol/transport errors
-
-          // For SSE, we should at least get some events
-          expect(events.length).toBeGreaterThan(0);
-
-          // Look for any message event (the SSE connection is working)
-          const messageEvent = events.find((event: any) => event.type === 'message');
-          
-          // If we get a message event, the SSE functionality is working
-          // The specific tool response format may vary with the official SDK
-          if (messageEvent) {
-            expect(messageEvent.type).toBe('message');
-            // SSE functionality is confirmed working
-          } else {
-            // If no message events, at least verify we got the SSE connection
-            expect(events.length).toBeGreaterThanOrEqual(0);
-          }
-
-          callback(null, null);
-        } catch (assertionError) {
-          callback(assertionError, null);
-        }
+      .set({
+        'Content-Type': 'application/json',
+        Accept: 'application/json',
+        'mcp-session-id': mcpSessionId,
       })
-      .end(done); // Use Jest's done callback for async test completion
+      .send({
+        jsonrpc: '2.0',
+        id: 'test_after_termination',
+        method: 'tools/list',
+        params: {},
+      })
+      .expect(400);
+
+    expect(testResponse.text).toContain('Invalid or missing session ID');
+
+    // Re-initialize session for cleanup
+    await initializeMcpSession();
   });
 
-  describe('Entity CRUD via /mcp (JSON)', () => {
-    it('T_HTTPSTREAM_CRUD_add-component: should add a primary component', async () => {
-      sharedTestComponentId = `e2e-http-crud-comp-${Date.now()}`;
+  describe('Entity CRUD via HTTP Streaming', () => {
+    it('T_HTTP_MODERN_CRUD_add-component: should add a component', async () => {
+      sharedTestComponentId = `modern-crud-comp-${Date.now()}`;
       const compArgs = {
         repository: testRepository,
         branch: testBranch,
         id: sharedTestComponentId,
-        name: 'HTTP CRUD Primary',
+        name: 'Modern CRUD Primary',
         kind: 'module',
         status: 'active' as ComponentStatus,
         clientProjectRoot: clientProjectRootForTest,
       };
       const payload = {
         jsonrpc: '2.0',
-        id: 'add_comp_crud',
+        id: 'add_comp_modern_crud',
         method: 'tools/call',
         params: { name: 'add-component', arguments: compArgs },
       };
-      
-      const response = await makeMcpRequest(payload);
-      expect(response.body.id).toBe('add_comp_crud');
-      expect(response.body.result).toBeDefined();
-      expect(response.body.result.content).toBeDefined();
-      expect(response.body.result.content[0]).toBeDefined();
 
-      // Parse the JSON result from the MCP SDK format
+      const response = await makeMcpRequest(payload);
+      expect(response.body.id).toBe('add_comp_modern_crud');
+      expect(response.body.result).toBeDefined();
+
       const toolResult = JSON.parse(response.body.result.content[0].text);
       expect(toolResult.success).toBe(true);
       expect(toolResult.message).toContain(sharedTestComponentId);
     });
 
-    it('T_HTTPSTREAM_CRUD_add-component-dependent: should add a dependent component', async () => {
-      expect(sharedTestComponentId).not.toBeNull();
-      sharedDependentComponentId = `e2e-http-crud-dep-${Date.now()}`;
-      const compArgs = {
-        repository: testRepository,
-        branch: testBranch,
-        id: sharedDependentComponentId,
-        name: 'HTTP CRUD Dependent',
-        kind: 'service',
-        status: 'active' as ComponentStatus,
-        depends_on: [sharedTestComponentId!],
-        clientProjectRoot: clientProjectRootForTest,
-      };
-      const payload = {
-        jsonrpc: '2.0',
-        id: 'add_dep_crud',
-        method: 'tools/call',
-        params: { name: 'add-component', arguments: compArgs },
-      };
-      
-      const response = await makeMcpRequest(payload);
-      expect(response.body.id).toBe('add_dep_crud');
-      expect(response.body.result).toBeDefined();
-      expect(response.body.result.content).toBeDefined();
-      expect(response.body.result.content[0]).toBeDefined();
-
-      // Parse the JSON result from the MCP SDK format
-      const toolResult = JSON.parse(response.body.result.content[0].text);
-      expect(toolResult.success).toBe(true);
-    });
-
-    it('T_HTTPSTREAM_CRUD_add-decision: should add a decision', async () => {
-      sharedTestDecisionId = `e2e-http-crud-dec-${Date.now()}`;
-      const decisionArgs = {
-        repository: testRepository,
-        branch: testBranch,
-        id: sharedTestDecisionId,
-        name: 'HTTP CRUD Decision',
-        date: '2024-01-01',
-        context: 'Test decision',
-        clientProjectRoot: clientProjectRootForTest,
-      };
-      const payload = {
-        jsonrpc: '2.0',
-        id: 'add_dec_crud',
-        method: 'tools/call',
-        params: { name: 'add-decision', arguments: decisionArgs },
-      };
-      
-      const response = await makeMcpRequest(payload);
-      expect(response.body.id).toBe('add_dec_crud');
-      expect(response.body.result).toBeDefined();
-      expect(response.body.result.content).toBeDefined();
-      expect(response.body.result.content[0]).toBeDefined();
-
-      // Parse the JSON result from the MCP SDK format
-      const toolResult = JSON.parse(response.body.result.content[0].text);
-      expect(toolResult.success).toBe(true);
-    });
-
-    it('T_HTTPSTREAM_CRUD_add-rule: should add a rule', async () => {
-      sharedTestRuleId = `e2e-http-crud-rule-${Date.now()}`;
-      const ruleArgs = {
-        repository: testRepository,
-        branch: testBranch,
-        id: sharedTestRuleId,
-        name: 'HTTP CRUD Rule',
-        created: '2024-01-01',
-        content: 'Test rule',
-        status: 'active' as const,
-        clientProjectRoot: clientProjectRootForTest,
-      };
-      const payload = {
-        jsonrpc: '2.0',
-        id: 'add_rule_crud',
-        method: 'tools/call',
-        params: { name: 'add-rule', arguments: ruleArgs },
-      };
-      
-      const response = await makeMcpRequest(payload);
-      expect(response.body.id).toBe('add_rule_crud');
-      expect(response.body.result).toBeDefined();
-      expect(response.body.result.content).toBeDefined();
-      expect(response.body.result.content[0]).toBeDefined();
-
-      // Parse the JSON result from the MCP SDK format
-      const toolResult = JSON.parse(response.body.result.content[0].text);
-      expect(toolResult.success).toBe(true);
-    });
-  });
-
-  describe('Traversal and Graph Tools via /mcp (JSON responses)', () => {
-    // Assumes components from CRUD tests or beforeAll are available (sharedTestComponentId, sharedDependentComponentId)
-    it('T_HTTPSTREAM_JSON_get-component-dependencies: should retrieve dependencies', async () => {
-      expect(sharedDependentComponentId).toBeDefined();
+    it('T_HTTP_MODERN_CRUD_get-dependencies: should retrieve dependencies', async () => {
       const toolArgs = {
         repository: testRepository,
         branch: testBranch,
-        componentId: sharedDependentComponentId,
+        componentId: dependentComponentId,
         clientProjectRoot: clientProjectRootForTest,
       };
       const payload = {
         jsonrpc: '2.0',
-        id: 'get_deps_json',
+        id: 'get_deps_modern',
         method: 'tools/call',
         params: { name: 'get-component-dependencies', arguments: toolArgs },
       };
-      
-      const response = await makeMcpRequest(payload);
-      expect(response.body.id).toBe('get_deps_json');
-      expect(response.body.result).toBeDefined();
-      expect(response.body.result.content).toBeDefined();
-      expect(response.body.result.content[0]).toBeDefined();
 
-      // Parse the JSON result from the MCP SDK format
+      const response = await makeMcpRequest(payload);
+      expect(response.body.id).toBe('get_deps_modern');
+      expect(response.body.result).toBeDefined();
+
       const resultWrapper = JSON.parse(response.body.result.content[0].text);
       expect(resultWrapper.status).toBe('complete');
       expect(Array.isArray(resultWrapper.dependencies)).toBe(true);
       expect(
-        resultWrapper.dependencies.some((c: Component) => c.id === sharedTestComponentId),
+        resultWrapper.dependencies.some((c: Component) => c.id === testComponentId),
       ).toBe(true);
-    });
-
-    it('T_HTTPSTREAM_JSON_shortest-path: should find a shortest path', async () => {
-      expect(sharedDependentComponentId).toBeDefined();
-      const toolArgs = {
-        repository: testRepository,
-        branch: testBranch,
-        startNodeId: sharedDependentComponentId,
-        endNodeId: sharedTestComponentId!,
-        relationshipTypes: ['DEPENDS_ON'],
-        direction: 'OUTGOING',
-        clientProjectRoot: clientProjectRootForTest,
-        projectedGraphName: 'sp_json_test_graph',
-        nodeTableNames: ['Component'],
-        relationshipTableNames: ['DEPENDS_ON'],
-      };
-      const payload = {
-        jsonrpc: '2.0',
-        id: 'sp_json',
-        method: 'tools/call',
-        params: { name: 'shortest-path', arguments: toolArgs },
-      };
-      
-      const response = await makeMcpRequest(payload);
-      expect(response.body.id).toBe('sp_json');
-      expect(response.body.result).toBeDefined();
-      expect(response.body.result.content).toBeDefined();
-      expect(response.body.result.content[0]).toBeDefined();
-
-      // Parse the JSON result from the MCP SDK format
-      const resultWrapper = JSON.parse(response.body.result.content[0].text);
-      expect(resultWrapper.status).toBe('complete');
-      expect(resultWrapper.results.pathFound).toBe(true);
-      expect(Array.isArray(resultWrapper.results.path)).toBe(true);
-      expect(resultWrapper.results.path.length).toBeGreaterThanOrEqual(1);
-      expect(resultWrapper.results.path[0].id).toBe(sharedDependentComponentId);
-      if (resultWrapper.results.path.length >= 2) {
-        expect(resultWrapper.results.path[resultWrapper.results.path.length - 1].id).toBe(
-          sharedTestComponentId,
-        );
-      }
-    });
-
-    it('T_HTTPSTREAM_JSON_get-component-dependents: should retrieve dependents', async () => {
-      expect(sharedTestComponentId).toBeDefined();
-      expect(sharedDependentComponentId).toBeDefined();
-      const toolArgs = {
-        repository: testRepository,
-        branch: testBranch,
-        componentId: sharedTestComponentId!,
-        clientProjectRoot: clientProjectRootForTest,
-      };
-      const payload = {
-        jsonrpc: '2.0',
-        id: 'get_dependents_json',
-        method: 'tools/call',
-        params: { name: 'get-component-dependents', arguments: toolArgs },
-      };
-      
-      const response = await makeMcpRequest(payload);
-      expect(response.body.id).toBe('get_dependents_json');
-      expect(response.body.result).toBeDefined();
-      expect(response.body.result.content).toBeDefined();
-      expect(response.body.result.content[0]).toBeDefined();
-
-      // Parse the JSON result from the MCP SDK format
-      const resultWrapper = JSON.parse(response.body.result.content[0].text);
-      expect(resultWrapper.status).toBe('complete');
-      expect(Array.isArray(resultWrapper.dependents)).toBe(true);
-      // Check if the dependent component we created is listed
-      expect(
-        resultWrapper.dependents.some((c: Component) => c.id === sharedDependentComponentId),
-      ).toBe(true);
-    });
-
-    it('T_HTTPSTREAM_JSON_shortest-path_reflexive: should handle reflexive shortest path', async () => {
-      expect(sharedTestComponentId).toBeDefined();
-      const toolArgs = {
-        repository: testRepository,
-        branch: testBranch,
-        startNodeId: sharedTestComponentId!,
-        endNodeId: sharedTestComponentId!,
-        clientProjectRoot: clientProjectRootForTest,
-        projectedGraphName: 'sp_reflexive_json_test_graph',
-        nodeTableNames: ['Component'],
-        relationshipTableNames: [],
-      };
-      const payload = {
-        jsonrpc: '2.0',
-        id: 'sp_reflex_json',
-        method: 'tools/call',
-        params: { name: 'shortest-path', arguments: toolArgs },
-      };
-      
-      const response = await makeMcpRequest(payload);
-      expect(response.body.id).toBe('sp_reflex_json');
-      expect(response.body.result).toBeDefined();
-      expect(response.body.result.content).toBeDefined();
-      expect(response.body.result.content[0]).toBeDefined();
-
-      // Parse the JSON result from the MCP SDK format
-      const resultWrapper = JSON.parse(response.body.result.content[0].text);
-      expect(resultWrapper.status).toBe('complete');
-      expect(resultWrapper.results).toBeDefined(); // Ensure results object exists
-      expect(resultWrapper.results.pathFound).toBe(false);
-      expect(Array.isArray(resultWrapper.results.path)).toBe(true);
-      expect(resultWrapper.results.path.length).toBe(0);
     });
   });
 
-  describe('Algorithm Tools via /mcp (JSON responses)', () => {
-    const algorithmTestCases = [
-      {
-        name: 'k-core-decomposition',
-        args: {
-          k: 1,
-          projectedGraphName: 'kcore_json_test_graph',
-          nodeTableNames: ['Component'],
-          relationshipTableNames: ['DEPENDS_ON'],
-        },
-        expectedDataKeyInWrapper: 'results',
-        checkField: 'components',
-      },
-      {
-        name: 'louvain-community-detection',
-        args: {
-          projectedGraphName: 'louvain_json_test_graph',
-          nodeTableNames: ['Component'],
-          relationshipTableNames: ['DEPENDS_ON'],
-        },
-        expectedDataKeyInWrapper: 'results',
-        checkModularity: true,
-        checkField: 'communities',
-      },
-      {
-        name: 'pagerank',
-        args: {
-          projectedGraphName: 'pagerank_json_test_graph',
-          nodeTableNames: ['Component'],
-          relationshipTableNames: ['DEPENDS_ON'],
-        },
-        expectedDataKeyInWrapper: 'results',
-        checkField: 'ranks',
-      },
-      {
-        name: 'strongly-connected-components',
-        args: {
-          projectedGraphName: 'scc_json_test_graph',
-          nodeTableNames: ['Component'],
-          relationshipTableNames: ['DEPENDS_ON'],
-        },
-        expectedDataKeyInWrapper: 'results',
-        checkField: 'components',
-      },
-      {
-        name: 'weakly-connected-components',
-        args: {
-          projectedGraphName: 'wcc_json_test_graph',
-          nodeTableNames: ['Component'],
-          relationshipTableNames: ['DEPENDS_ON'],
-        },
-        expectedDataKeyInWrapper: 'results',
-        checkField: 'components',
-      },
-    ];
-
-    for (const toolSetup of algorithmTestCases) {
-      it(`T_HTTPSTREAM_JSON_ALGO_${toolSetup.name}: should execute and return wrapper`, async () => {
-        const toolArgs = {
-          repository: testRepository,
-          branch: testBranch,
-          ...toolSetup.args,
-          clientProjectRoot: clientProjectRootForTest,
-        };
+  describe('Session Management Features', () => {
+    it('T_HTTP_MODERN_SESSION_001: Should track multiple concurrent requests', async () => {
+      const requests = [];
+      for (let i = 0; i < 3; i++) {
         const payload = {
           jsonrpc: '2.0',
-          id: `algo_json_${toolSetup.name}`,
+          id: `concurrent_request_${i}`,
           method: 'tools/call',
-          params: { name: toolSetup.name, arguments: toolArgs },
+          params: {
+            name: 'get-metadata',
+            arguments: {
+              repository: testRepository,
+              branch: testBranch,
+              clientProjectRoot: clientProjectRootForTest,
+            },
+          },
         };
-        
-        const response = await makeMcpRequest(payload);
-        expect(response.body.id).toBe(`algo_json_${toolSetup.name}`);
-        expect(response.body.result).toBeDefined();
-        expect(response.body.result.content).toBeDefined();
-        expect(response.body.result.content[0]).toBeDefined();
+        requests.push(makeMcpRequest(payload));
+      }
 
-        // Parse the JSON result from the MCP SDK format
-        const resultWrapper = JSON.parse(response.body.result.content[0].text);
-        expect(resultWrapper).toBeDefined();
-        expect(resultWrapper.status).toBe('complete');
+      const responses = await Promise.all(requests);
+      for (let i = 0; i < 3; i++) {
+        expect(responses[i].body.id).toBe(`concurrent_request_${i}`);
+        expect(responses[i].body.result).toBeDefined();
+      }
+    });
 
-        const dataContainer = resultWrapper[toolSetup.expectedDataKeyInWrapper]; // This is now resultWrapper.results
-        expect(dataContainer).toBeDefined();
+    it('T_HTTP_MODERN_SESSION_002: Should handle GET request for SSE notifications', async () => {
+      // Test that the SSE endpoint is available but may not have content yet
+      const response = await request(BASE_URL)
+        .get('/mcp')
+        .set('mcp-session-id', mcpSessionId)
+        .set('Accept', 'text/event-stream');
 
-        // Updated checks to look inside dataContainer (which is resultWrapper.results)
-        if (toolSetup.checkField) {
-          expect(dataContainer[toolSetup.checkField]).toBeDefined();
-          expect(Array.isArray(dataContainer[toolSetup.checkField])).toBe(true);
-        }
-
-        if (toolSetup.checkModularity) {
-          // Modularity for Louvain is returned at the same level as 'communities' within the 'results' object from the operation
-          expect(dataContainer).toHaveProperty('modularity');
-        }
-      });
-    }
+      // For modern HTTP streaming, this should either work or return 405
+      expect([200, 405]).toContain(response.status);
+    });
   });
 });
